Split App helpers into editor modules and hooks
